Migrate enabled/disabled features test to unit test
