+++ app/BUILD.gn
@@ -96,71 +96,80 @@ source_set("browser_tests") {
       "//brave/components/commands/common",
       "//chrome/app:command_ids",
       "//chrome/browser",
-      "//chrome/browser/apps/link_capturing:link_capturing",
-      "//chrome/browser/companion/core",
-      "//chrome/browser/enterprise/connectors/analysis:features",
       "//chrome/browser/promos:utils",
       "//chrome/browser/ui",
-      "//chrome/common:chrome_features",
       "//chrome/common:non_code_constants",
-      "//chrome/common/companion/visual_query:flags",
       "//chrome/test",
       "//chrome/test:test_support",
       "//chrome/test:test_support_ui",
-      "//components/aggregation_service",
-      "//components/attribution_reporting",
-      "//components/autofill/core/common",
-      "//components/commerce/core:feature_list",
       "//components/component_updater",
-      "//components/content_settings/core/common:features",
       "//components/embedder_support",
-      "//components/heap_profiling/in_process",
-      "//components/history_clusters/core",
-      "//components/language/core/common",
-      "//components/lens",
-      "//components/manta",
-      "//components/metrics/structured:structured_metrics_features",
-      "//components/network_time",
-      "//components/omnibox/browser",
-      "//components/optimization_guide/core",
-      "//components/page_image_service",
-      "//components/page_info/core",
-      "//components/performance_manager",
-      "//components/permissions",
-      "//components/privacy_sandbox",
-      "//components/reading_list/features:flags",
-      "//components/safe_browsing/core/common",
-      "//components/search",
-      "//components/segmentation_platform/public",
-      "//components/send_tab_to_self",
-      "//components/signin/public/base:signin_buildflags",
-      "//components/subresource_filter/core/common",
-      "//components/translate/core/common",
-      "//extensions/common",
-      "//services/device/public/cpp:device_features",
       "//testing/gtest",
-      "//ui/accessibility:ax_base",
     ]
-
-    if (is_linux || is_win || is_mac) {
-      deps += [ "//components/device_signals/core/common:features" ]
-    }
   }
 }
 
 source_set("unit_tests") {
   testonly = true
-
-  if (!is_android && !is_ios) {
-    sources = [ "command_utils_unittest.cc" ]
-    deps = [
+  sources = [ "feature_defaults_unittest.cc" ]
+  if (!is_android) {
+    sources += [ "command_utils_unittest.cc" ]
+  }
+  deps = [
+    "//base",
+    "//base/test:test_support",
+    "//chrome/browser",
+    "//chrome/browser/companion/core",
+    "//chrome/browser/enterprise/connectors/analysis:features",
+    "//chrome/browser/ui",
+    "//chrome/common/companion/visual_query:flags",
+    "//components/aggregation_service",
+    "//components/attribution_reporting",
+    "//components/autofill/core/common",
+    "//components/commerce/core:feature_list",
+    "//components/content_settings/core/common:features",
+    "//components/device_signals/core/common:features",
+    "//components/heap_profiling/in_process",
+    "//components/history_clusters/core",
+    "//components/language/core/common",
+    "//components/lens",
+    "//components/metrics/structured:structured_metrics_features",
+    "//components/network_time",
+    "//components/omnibox/browser",
+    "//components/optimization_guide/core",
+    "//components/page_image_service",
+    "//components/page_info/core",
+    "//components/performance_manager",
+    "//components/permissions",
+    "//components/privacy_sandbox",
+    "//components/safe_browsing/core/common",
+    "//components/search",
+    "//components/segmentation_platform/public",
+    "//components/shared_highlighting/core/common",
+    "//components/signin/public/base:signin_buildflags",
+    "//components/subresource_filter/core/common",
+    "//gpu/config",
+    "//services/device/public/cpp:device_features",
+    "//services/network/public/cpp",
+    "//testing/gtest",
+    "//ui/base",
+  ]
+  if (is_android) {
+    deps += [
+      "//android_webview/common",
+      "//components/password_manager/core/common:features",
+    ]
+  } else {
+    deps += [
       ":command_ids",
-      "//base",
-      "//base/test:test_support",
       "//brave/components/commands/common",
       "//chrome/app:command_ids",
-      "//chrome/browser/ui",
-      "//testing/gtest",
+      "//chrome/browser/apps/link_capturing:link_capturing",
+      "//chrome/common:chrome_features",
+      "//components/manta",
+      "//components/translate/core/common",
+      "//extensions/common",
+      "//ui/accessibility:ax_base",
     ]
   }
 }
